python-scripts/kafka-scripts/requirements.txt
requests
kafka-python
python-dotenv
pandas
asyncio
aiohttp